Roulette

See merge request it1901/groups-2021/gr2124/gr2124!10

+++ modules-template/core/src/main/java/roulette/ListGuess.java
@@ -2,7 +2,7 @@ package roulette;
 
 public class ListGuess extends Guess {
 
-	public ListGuess(int amount, int start, int end) {
+	public ListGuess(double amount, int start, int end) {
 		super(amount);
 
 		for (int i = start; i <= end; i++) {
